+++ services/datasources/lib/datasources/search/twitter.rb
@@ -579,13 +579,13 @@ module CartoDB
         end
 
         # Call this inside specs to override returned class
-        # @param override_class SearchTweet|nil (optional)
-        # @return SearchTweet
+        # @param override_class Carto::SearchTweet|nil (optional)
+        # @return Carto::SearchTweet
         def audit_entry(override_class = nil)
           if @audit_entry.nil?
             if override_class.nil?
-              require_relative '../../../../../app/models/search_tweet'
-              @audit_entry = ::SearchTweet.new
+              require_relative '../../../../../app/models/carto/search_tweet'
+              @audit_entry = Carto::SearchTweet.new
             else
               @audit_entry = override_class.new
             end