+++ spec/requests/superadmin/feature_flag_spec.rb
@@ -1,56 +1,58 @@
 require_relative '../../acceptance_helper'
 
-describe FeatureFlag do
+describe Carto::FeatureFlag do
+  let(:feature_flag) { create(:feature_flag) }
+
   describe '#create' do
-    it 'should create feature flag' do
-      feature_flag = FactoryGirl.build(:feature_flag)
+    let(:feature_flag) { build(:feature_flag) }
 
-      expect {
-        post superadmin_feature_flags_url, { feature_flag: feature_flag }.to_json, superadmin_headers
+    context 'when everything is OK' do
+      it 'should create feature flag' do
+        payload = { feature_flag: feature_flag.attributes }.to_json
+        expect {
+          post superadmin_feature_flags_url, payload, superadmin_headers
+        }.to change(Carto::FeatureFlag, :count).by(1)
 
-        response.status.should == 204
-      }.to change(FeatureFlag, :count).by(1)
+        expect(response.status).to eq(204)
+      end
+    end
+
+    context 'when an error occurs' do
+      it 'returns an error' do
+        post superadmin_feature_flags_url, {}.to_json, superadmin_headers
+
+        expect(response.status).to eq(500)
+      end
     end
   end
 
   describe '#update' do
     it 'should update feature flag name' do
-      feature_flag      = FactoryGirl.build(:feature_flag)
-      test_feature_flag = FactoryGirl.build(:feature_flag)
+      payload = { feature_flag: feature_flag.attributes.merge('name' => 'new_name') }.to_json
+      put superadmin_feature_flag_url(feature_flag.id), payload, superadmin_headers
 
-      test_feature_flag.id   = feature_flag.id
-      test_feature_flag.name = "test_new_name"
-      test_feature_flag.save
-
-      old_name = feature_flag.name
-
-      expect {
-        put superadmin_feature_flag_url(feature_flag.id), { feature_flag: test_feature_flag }.to_json, superadmin_headers
-
-        feature_flag.reload
-      }.to change(feature_flag, :name).from(old_name).to(test_feature_flag.name)
+      expect(feature_flag.reload.name).to eq('new_name')
     end
   end
 
   describe '#destroy' do
+    let(:payload) { {}.to_json }
+    let(:user) { create(:user) }
+
+    before { feature_flag.save! }
 
     it 'should destroy feature flag' do
-      feature_flag = FactoryGirl.create(:feature_flag)
-
       expect {
-        delete superadmin_feature_flag_url(feature_flag.id), { feature_flag: feature_flag }.to_json, superadmin_headers
-      }.to change(FeatureFlag, :count).by(-1)
+        delete superadmin_feature_flag_url(feature_flag.id), payload, superadmin_headers
+      }.to change(Carto::FeatureFlag, :count).by(-1)
     end
 
     it 'should destroy feature flag user relations' do
-      feature_flag = FactoryGirl.create(:feature_flag)
-      user         = FactoryGirl.create(:user)
-
-      feature_flag_user = FactoryGirl.create(:feature_flags_user, feature_flag_id: feature_flag.id, user_id: user.id)
+      user.activate_feature_flag!(feature_flag)
 
       expect {
-        delete superadmin_feature_flag_url(feature_flag.id), { feature_flag: feature_flag }.to_json, superadmin_headers
-      }.to change(FeatureFlagsUser, :count).by(-1)
+        delete superadmin_feature_flag_url(feature_flag.id), payload, superadmin_headers
+      }.to change(Carto::FeatureFlagsUser, :count).by(-1)
     end
   end